Issue #7: Add organisation validation
When importing organisations with its hierarchy, there are some
requirements in order to make the hiearchy be imported properly. If we
do not have organisation code or the organisation node has been
previously flattened it doesn't make sense to form an hierarchy where
organisation belongs to a single place.

This change will skip organisations in the hierarchy along with its
children when:
  a) organisation has no code
  b) organisation has been already processed